Have concourse worker use any locally configured proxies.
If a proxy is configured, unfortunately Upstart will not start processes
with /etc/environment variables available. This means that if you have
an external proxy setup (e.g. via vagrant-proxyconf) then things will
not work.

Use the workaround listed at: https://serverfault.com/a/717057

This will explicity set the three parameters:

 --http-proxy
 --https-proxy
 --no-proxy

But if there is no value specified in /etc/environment
this has no ill effects.

The configuration will be passed down to the containers and they will
also utilise the passed through proxy as well.